Address most PR comments.
Update Podfile.lock
