+++ Riot/Managers/Analytics/Helpers/MXTaskProfileName+Analytics.swift
@@ -1,0 +1,42 @@
+// 
+// Copyright 2021 New Vector Ltd
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+
+import AnalyticsEvents
+
+extension MXTaskProfileName {
+    var analyticsName: AnalyticsEvent.PerformanceTimer.Name? {
+        switch self {
+        case .startupIncrementalSync:
+            return .StartupIncrementalSync
+        case .startupInitialSync:
+            return .StartupInitialSync
+        case .startupLaunchScreen:
+            return .StartupLaunchScreen
+        case .startupStorePreload:
+            return .StartupStorePreload
+        case .startupMountData:
+            return .StartupStoreReady
+        case .initialSyncRequest:
+            return .InitialSyncRequest
+        case .initialSyncParsing:
+            return .InitialSyncParsing
+        case .notificationsOpenEvent:
+            return .NotificationsOpenEvent
+        default:
+            return nil
+        }
+    }
+}
