+++ Riot/Modules/TabBar/TabBarCoordinator.swift
@@ -491,6 +491,10 @@ final class TabBarCoordinator: NSObject, TabBarCoordinatorType {
     private func presentAnalyticsPrompt(with session: MXSession) {
         let parameters = AnalyticsPromptCoordinatorParameters(session: session, navigationRouter: navigationRouter)
         let coordinator = AnalyticsPromptCoordinator(parameters: parameters)
+        coordinator.completion = { [weak self] in
+            self?.remove(childCoordinator: coordinator)
+        }
+        
         coordinator.start()
         add(childCoordinator: coordinator)
     }
