+++ Riot/SupportingFiles/Riot-Bridging-Header.h
@@ -45,6 +45,7 @@
 #import "RoomInputToolbarView.h"
 #import "NSArray+Element.h"
 #import "ShareItemSender.h"
+#import "HTMLFormatter.h"
 
 // MatrixKit common imports, shared with all targets
 #import "MatrixKit-Bridging-Header.h"
